use volume's hash algorithm
